refactor(minesweeper): BoardBuilder를 dsl 패키지로 이동
